Refactor: split common decorator class from the schema implementation
